tools/Info-Fetcher/requirements.txt
requests>=2.28.0
pandas>=1.5.0
streamlit>=1.24.0
plotly>=5.14.0
xlsxwriter>=3.0.9
streamlit-echarts>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
